Refactor avro definitions out of container package
