+++ container/avro/block.avsc
