+++ container/avro/generate.go
@@ -1,0 +1,3 @@
+package avro
+
+//go:generate $GOPATH/bin/gogen-avro --package avro . block.avsc header.avsc
