--- container/generate.go
@@ -1,3 +1,0 @@
-package container
-
-//go:generate $GOPATH/bin/gogen-avro --package container . block.avsc header.avsc
